Move selectLabelText to shared utility
