+++ packages/block-library/src/navigation-link/shared/index.js
@@ -17,3 +17,4 @@ export { useIsInvalidLink } from './use-is-invalid-link';
 export { InvalidDraftDisplay } from './invalid-draft-display';
 export { useEnableLinkStatusValidation } from './use-enable-link-status-validation';
 export { useIsDraggingWithin } from './use-is-dragging-within';
+export { selectLabelText } from './select-label-text';
